Allow specifying rpms in building helm chart
This commit adds support for manually specifying rpms when building helm charts.
This is specified with a --rpm option that takes a comma-separated list of rpms
providing manifests for building the helm application tarball
These are then merged to generate APP_NAME.yaml, which is included in the
application tarball.

Change-Id: I8bbc5a06a369cfe5db5ed3667d76d7d44d5b52e3
Story: 2005350
Task: 33526
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>