Update mockchain-parallel wrapper for mock-1.4.20
The recently released mock-1.4.20 replaces mockchain with a bash
wrapper that calls mock --chain. However, the StarlingX
mockchain-parallel utility assumes mockchain is python, and fails to
determine the mockchain version as a result. This causes build
failures when mock-1.4.20 is installed.

This update enhances mockchain-parallel to use the version from the
installed mock rpm if the version cannot be determined from mockchain
itself. This allows mockchain-parallel to work with mock-1.4.20, as
well as previously supported versions of mock.

Change-Id: Ic0360fc64e277f1816036b6259e6e51d94c7f368
Closes-Bug: 1849147
Signed-off-by: Don Penney <don.penney@windriver.com>
(cherry picked from commit 639e435594d213709df09fcf1fb10f2d03c55094)