Build avoidance fixes
1) Build avoidance fails on a clean workspace because directory
$MY_WORKSPACE/tmp doesn't exist when mktemp needs to
create a forth subdir under that path.

Add a 'mkdir -p $MY_WORKSPACE/tmp' to ensure it exists
prior to mktemp.

2) Failed to rebuild kmods after a kernel change and
building with build-avoidance enabled.  The problem was
a missing dependency cache that would have listed
the kmods as having dependencies on the kernel.
Reusing kmods from a prior build in conjunction with
a rebuilt kernel results int load with unusable kmods
due to a key mismatch.

Add the dependancy-cache to the list of artifacts to
syncronize as part of build avoidance.  A primed
dependency cache will allow build-pkgs to determine
that kmods must be rebuilt when the kernel has changed.

3) Make sure $MY_WORKSPACE/$BUILD_TYPE/repo
symlink exists after a build avoidance build.
If no packages require a rebuild, it was missing.

4) Fix an error message that was puzzling while debugging
the above issues.

Closes-Bug: 1864110
Change-Id: Ia912b2e3e110efbe69004d7978109ebb089d684b
Signed-off-by: Scott Little <scott.little@windriver.com>