Fix bug in stable-wheels.cfg
The previous commit for openstack/train upgrade modified the rjsmin
entry to download a prebuilt wheel file from pypi, rather than
building from a tar file. However, the type field was left as "tar"
rather than updating to "pypi", and the build fails as a result.

This update corrects that field.

Change-Id: I9ad77054f41925d3dfdd10889e0c1ef7683c711b
Story: 2006544
Task: 36618
Signed-off-by: Don Penney <don.penney@windriver.com>