Align stable-wheels.cfg with train constraints
This commit updates the stable-wheels.cfg to align with the
stable/train upper-constraints.txt file, to ensure the train images
built have the proper content:
- If an appropriate prebuilt wheel is available at pypi.org, remove
the entry from stable-wheels.cfg to avoid changing the configured
train constraint.
- If building from a tar or zip and the configured train constraint
has changed, update the entry with the new version.
- Added entres to build pyrsistent, psycopg2, and thriftpy2 from
tarballs, as they are needed for various images and no prebuilt wheels
are available.

Change-Id: I6cebb9a9867c37746500e3d6d4762ed7668d60df
Story: 2006544
Task: 37427
Signed-off-by: Don Penney <don.penney@windriver.com>