Add tag-management utility
This commit introduces a tag management utility for images to allow
retagging of specific images. This allows the StarlingX ansible
playbooks and packaged charts to use a static image tag rather than the
moving stable-latest tag.

This utility can only be run by a StarlingX build team member with
publishing capability on the docker hub, to allow for updated tags to
be available publicly, or as part of a CENGN build job.

Change-Id: I7659695c701ccf3bfe8255b24ad0b1896b59cc23
Partial-Bug: 1854869
Signed-off-by: Don Penney <don.penney@windriver.com>