Lock down upper-constraints for wheels
Currently, the images built via loci are python2-based. Given that
support for python2 has ended, most new module versions in pypi are
python3-only. As a result, these new versions will fail to install in
the image during the build.

A recent change to the upstream upper-constraints upversioned the
ovsdbapp module to a python3-only version, which caused the stx-nova
and stx-neutron image builds to fail. As a temporary workaround until
we can move these images to python3, this update locks down the
upper-constraints and global-requirements files to a version prior to
this change.

Change-Id: Iaf0e9c76c96b21c398345dfc6b0b8715e978e879
Closes-Bug: 1863957
Signed-off-by: Don Penney <don.penney@windriver.com>