Fix potential infinite loop in build-iso
Function generate_dep_list is creating a DEPLISTFILE
with a lot of unwanted 'installing package xxx' messages
when it should only contain a list of unresolved
dependencies.

On the next iteration we waste a lot of time trying to install
nonexistant packages 'installing' and 'packages', and re-installing
packages we already have.

Instead of converging on an empty DEPLISTFILE, we converge
on a DEPLISTFILE that just reiterates the same list of
'installing package' if we are lucky, which we usually are,
but not always.

Closes-Bug: 1877483
Change-Id: I971725558eb628c6d6b5d0b42d0bcab6b46e722b
Signed-off-by: Scott Little <scott.little@windriver.com>