Add subdir PKG_GITREVCOUNT support
This update introduces the PKG_GITREVCOUNT variable for package-based
auto-versioning. Similar to GITREVCOUNT, PKG_GITREVCOUNT will
calculate the number of commits in the corresponding package's
directory. This allows each package within a git repo to be
independently auto-versioned.

Use of the PKG_BASE_SRCREV variable is optional. If it is not defined,
the PKG_GITREVCOUNT will count all commits in the package directory.
If it is defined, it will count a range of commits, with
PKG_BASE_SRCREV as the starting point.

To use this, include the following in the package's build_srpm.data
file:
TIS_PATCH_VER=PKG_GITREVCOUNT

This can also be combined with GITREVCOUNT, if a package has local
files in addition to an external repo to be factored in. This would
add the count from PKG_GITREVCOUNT to GITREVCOUNT in calculating the
TIS_PATCH_VER value:
TIS_PATCH_VER=GITREVCOUNT+PKG_GITREVCOUNT

Change-Id: Iab3a8c8a212d292f26d5418a881de69566758134
Story: 2006166
Task: 39765
Signed-off-by: Don Penney <don.penney@windriver.com>