build-helm-charts.sh support for multi-line url substitution
One of the steps under build-helm-charts.sh is to modify
upstream manifests by substituting docker image locations and tags
with ones that reference the images we have built and published.

Some newer manifests we would like to include are now using a
multi-line format that isn't supported by the current script.

Examples would be ...
     image: <host>[:<port>]/<path>/<image-name>
     imageTag: <tag>

or
     image:
        repository: <host>[:<port>]/<path>/<image-name>
        tag: <tag>

Our fix is to abandon sed, and substitute a routine
that walks the data structures looking substitutions
in either one or two line format.

Closes-Bug: 1893261
Depends-On: https://review.opendev.org/750710
Change-Id: I81be885e0c2460be9319fe403ef8fe938fe35549
Signed-off-by: Scott Little <scott.little@windriver.com>