Fix update-efiboot-image loop device leak
If update-efiboot-image exits abnormally or is terminated
from an external signal, it has been observed to leave
dangling mounts and/or fails to release the loop device.
Over time, all loop devices can be consumed on the build host.

This update adds an exit handler to ensure release of the
mount and loop device.

Closes-Bug: 1896822
Change-Id: I1891df134ae583750ca1d4858bdc2402e6a67ff6
Signed-off-by: Scott Little <scott.little@windriver.com>