Parse YAML with as YAML1.1 standard
Ruamel.yaml is YAML1.2 compatible by default. The YAML1.2 does not
support octals starts with 0 (the octals are required to start with
0o). For example, the octal number 0755 with be transferred into 755.
This commit set the ruamel.yaml load the filed as version 1.1. After
this commit, the octal number will be transferred to decimal. For
example, the octal 0755 will be re-write to 493. And it will be
load/dump correctly in the following steps.

Closes-Bug: 1896530
Change-Id: If8be9d8d4b22ccce99307b5388313ba85a71be74
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>