Remove Django from wheels tarball
Wheels added to the wheels.cfg file result in modification to the
upper-constraints.txt included in the wheels tarball to point to the
explicit wheel added by the wheels.cfg. The inclusion of Django
resulted in an upper-constraints.txt entry locking the version to
2.1.5.

However, the openstack/horizon requirements.txt file includes the
following: Django>=2.2,<3.0

This results in a conflict between the constraints and requirements.
Recent updates to the pip resolver exposed the conflict, and pip was
unable to find a version of Django to fit both.

Since the Django entry was just downloading a pre-built wheel, as
opposed to building an appropriate wheel from tarball, this update
removes the entry altogether to avoid creating this conflict.

Change-Id: I870a72f312c781a9b000fc98c4ed104c22469ebd
Closes-Bug: 1907290
Signed-off-by: Don Penney <don.penney@windriver.com>