build-img: Remove by-path symlinks for AWS images
In the graphical installation of an AWS-compatible image, there are
duplicate symlinks created by the udev persistent storage rules. This
change comments out the deprecated persistent storage rules in
/usr/lib/udev/rules.d/ during the setup work for the AWS image. This
issue has not been found in regular qcow2 images.

Tested on OpenStack in a distributed cloud system and on AWS.

Story: 2007858
Task: 41482

Change-Id: Icbf9e753042e29536bd92b13e4d779d97de1c5f2
Signed-off-by: Melissa Wang <melissa.wang@windriver.com>