Exclude .tox folder when checking file ages
The content of a .tox folder should not be considered when building
srpms. Additionally, as this folder can be fairly large and include
tens of thousands of files, a find listing can take some time. This
commit updates these find commands to exclude the .tox folder from the
search in order to save time if a user has run tox tests prior to the
build.

Change-Id: I035d730af8cdd443487f6ead3eec9e1e54f28709
Related-Bug: 1908940
Signed-off-by: Don Penney <don.penney@windriver.com>