build-img fails if not run under $MY_WORKSPACE
Build-img may need to generate a patch that is to be applied to the
image being produced.  The patch build tool drops the patch into
the current directory.  No explicit working directory was set.
A subsequent step tries to apply the patch, but assumes it is in
$MY_WORKSPACE.

We want all build artifacts to be placed under $MY_WORKSPACE, so
set $MY_WORKSPACE as the current directory before creating the patch.

Closes-Bug: 1915182
Signed-off-by: Scott Little <scott.little@windriver.com>
(cherry picked from commit 71d5ae9df9268be2803e2a45277e65c7c84c1fdb)
Change-Id: I8c34bc0c4697934cee2ca07334f817311c574872