Fix memory overcommit that caused OOM killer
Parallel package builds use large ramdisks.
It's important not to commit too much memory to these
ram disks, or we may push the system into memory exhaustion.
At that stage the Kernel will invoke the OOM killer,
It will likely select our build, or worse someone else's build,
to sacrifice.

The current algorithm only considers free memory at the instant
the parallel build starts.  It does not consider how many
other builds are in flight, but might not have allocated their
ramdisk yet.  The other build intends to use the memory, we see
the memory as free and try to use the same memory.

Solution is to consider total memory, and number of builds
already running or which might foreseeably start in the near future
(share factor) to derive an alternate estimate of memory available.
We then allocate the lesser amount.

Also fixed some issues with cleaning up of child processes when
a newer mockchain-parallel is in use.

Closes-Bug: 1917525
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: Iab178c6f9acbd5a209d66d0da21f367911f34905