Add MIRROR_LOCAL support to build-stx-images.sh
As the loci build runs in a container, it is unable to directly use a
PROJECT_REPO ref that requires ssh keys to be setup. In order to
support this, a MIRROR_LOCAL option is added to loci image directives
files.

If the image directives file has MIRROR_LOCAL=yes, the build utility
will first setup a bare clone of the PROJECT_REPO in the local
workspace and instead pass a reference to that to the loci build.

This functionality can be used by developers to test an image build
using a private forked repo that requires an ssh key to access, for
example. This should not be used in the formal build, as CENGN and
other developers are unlikely to have the required access.

Note: This requires that the build server in use has lighttpd server
setup with external access, as the container cannot access localhost.

Change-Id: Ibd260fbb47d2bece4dc27e1cf1c026fb5cd5ff0f
Story: 2003907
Task: 42364
Co-Authored-By: Scott Little <scott.little@windriver.com>
Signed-off-by: Don Penney <don.penney@windriver.com>