Fix wheels tarball generation
In d7c5a54ab94bce6635b83d91a807d28f97836a81, django was dropped in favor
of rfc3986. However the wheel was not added to the build-wheels
generation which breaks the docker images. Also add the migrate wheel
since it was mising as well. Add the required wheels in order to
build the docker image properly.

Test:
- Build new centos-stable-wheels tarball
- Build stx-keystone-api-proxy container

Closes-Bug: 1926795

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: Ib6f0abfdcc82ca14f92ebc5b45fe8df961e804ee