Three more improvements to branch creation tools.
1) Add default revision and soft/hard lockdown options to create_tags.sh.

These options already appear in the create_branches_and_tags.sh script.

Default revision set the default revision field of the manifest,
removing the need to specify a revision for each project.

Soft lockdown is content with tags already specified in the manifest.
Hard lockdown forces the use of sha's.

2) Improved determination of the local branch.

Need to work around inconsistencies of various git versions.
Also need to deal with 'repo' leaving us in detached head state and
local updates on a detached state, in which case we try
to find the 'best fit' local branch from which we diverged.

3) Ability to ask the branch/tag push scripts to bypass gerrit.

Successful use requires the git repositories and gerrit to
to permit direct push.  No true of the current opendev repo setup,
but might be useful for mirrors.

Closes-Bug: 1924762
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I2ba01e11ef27c1d10c816b5af6104afab440b3b3