Introduce platform based python package requirements
To enable the proper set up of tox dependencies that align with the
platform we introduce a location for managing those dependencies as part
of the build tools.

If we deem it necessary to stand up a formal requirements repo for
Starlingx (e.g. https://opendev.org/openstack/requirements) we can do so
in the future and migrate these files there.

Story: 2009101
Task: 43006
Change-Id: I78993a7109bdfdba2ae4050b06cd8d454e4404f2
Signed-off-by: Robert Church <robert.church@windriver.com>