Create new branch on bare clone for LOCI
This commit will force a branch creation on a bare clone of the locally
cloned repo so LOCI can use it afterwards to build images. The aim is to
prevent images from being build with code different from what is
checked-in on the local repo. That way we can avoid controlling branches
in two places (manifest and loci files).

Story: 2003907
Task: 42364
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I863ac450ddfaec3115f3ecbc2789fc5d7c9d887f