minimal_rpm_list.txt: ebtables->iptables-ebtables
This commit updates the minimal_rpm_list.txt file so that the ebtables
package included in the ISO image is replaced with iptables-ebtables.

Actually, given that iptables-ebtables obsoletes and replaces ebtables
in its RPM package metadata, this change is not truly necessary, because
iptables-ebtables is automatically installed instead of the legacy
ebtables package by the installer. The intent with this commit is to
ensure that the legacy ebtables package is not included in the ISO
image.

Verification:
- build-iso does not report any errors after having applied this change.
- Installation and bootstrap worked on a system with two controller
  hosts and two compute hosts.

Partial-Bug: #1949217
Depends-On: I63d557112c653d59b88ac3a4798dee0e89246612

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Change-Id: I6f99d0fe5e2632ea84f09d0a36680054c8c172f3