debrepack.py: support a tar ball with one and only file
The value cmdx returns isn't able to distinguish a tar ball
has one and only file with a tar ball has top directory. As
a result, the one and only file is removed by "--strip-components".

For example:
https://github.com/kubernetes-sigs/cri-tools/releases/download/v1.21.0/crictl-v1.21.0-linux-amd64.tar.gz

Also check the sub-directory to skip the tar ball with
one and only file.

Story: 2008846
Task: 43138

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: I985e1612411020176fa2f4f1314170e953d7d468