branching: Add ability to exclude projects from lock-down
Scripts create_branches_and_tags.sh and create_tags.sh
allow for automated manifest update, and offer lock-down
options to freeze all projects that weren't branched
or tagged on a specific SHA.

Sometimes we might not want to freeze ALL the other
projects.  Rather we need to control which are frozen
and which are not.  Freezing everything else remains
the preferred default behaviour.  Not freezing is an
exception.

This update adds options to exclude projects or remotes
from lock-down.

Closes-Bug: 1952473
Change-Id: I52d4b66ae44d06a6c05f492321fa2ae12fb3fc70
Signed-off-by: Scott Little <scott.little@windriver.com>