Provide a different source repo for spice-html5
We are noticing some intermittence on the gitlab service that provides
the spice-html5 source code defined on the default location on LOCI at
[1]. This commit creates the possibility to override the SPICE_REPO
location so we can decide for each image if we want to use a different
source and ref for spice-html5.

This patch also defaults to the github mirror of spice-html5 for all stx
images.

[1] https://github.com/openstack/loci/blob/master/Dockerfile#L29

TEST PLAN
PASS Build stx-nova image with SPICE_REPO and SPICE_REF changed
PASS Build images with a mix of changed and unchanged
     .stable_docker_images WRT spice-html5 source

Depends-On: https://review.opendev.org/c/openstack/loci/+/823577
Closes-Bug: #1956173
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I6872e42b08b2774666691b5dc4f0ff534160434a