build-avoidance: ensure we can write to a copied build environment.
A centos build using build avoidance will fail if the
upstream reference build is read only. The read only
is replicated in the user's build environment, and any
packages that need to be updates (since the reference
build) will fail with reason 'permission denied'

A read only reference build is rare, but can occur to
preserve important reference builds, such as a release build
or the last build prior tothe introduction of a major feature.

Solution is to add a chmod to ensure writability of copied content.

Closes-Bug: 1958156
Change-Id: Id7c4d9b90fe4b44cc8b1b351767a5dd74dc30e9a
Signed-off-by: Scott Little <scott.little@windriver.com>