Fixing default behavior of PYTHON3 for LOCI call
Recently, LOCI changed its default so images are python3 if nothing is
specified at the *docker_image files. With the recent upversion of LOCI
in the build tools [1], we need to explictly set PYTHON3=no to images
that are still using python2. As such, we need to pass WHEELS when no
PYTHON3 is provided and when PYTHON3 == "yes" (or != "no").

TEST PLAN
PASS Build image with PYTHON3=no (stx-pci-irq-affinity-agent)
PASS Build image with PYTHON3=yes (stx-cinder, modified manually)
PASS Build image with no PYTHON3 (stx-nova)

[1] https://opendev.org/starlingx/root/commit/00bd632a68b8172f5d4baec2b1e4f8a08b9d5be5

Closes-Bug: #1958696
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: Iec52b57e36da968d6076361b1b434234ef55d7e8