Fix downloader -s cengn-first strategy
downloader -s fails on setuptools, reason unknown.
That issue is being tracked separately.

However the point of the cengn-first strategy is to actually
try the upstream url if the download from cengn fails.

This update covers the Sourced download step of the below
launchpad.  The binary download step will be addressed as a
separate update.

Partial-Bug: 1967898
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: Ie216dc104022f3247931355e55d89bcbc375c401