patch-iso: Fix up paths
Commit c8c75e81b551 ("relocate /pxeboot to /var/pxeboot", 2021-12-15)
inadvertently introduced a few bugs to patch-iso that prevent successful
execution of this script when a patch with pxe-network-installer is
attempted to be used. This is necessary, for example, to be able to
patch kernel and initial RAM file system images in an ISO image.

Here is an example for the error messages:

  + cp --preserve=all pxeboot/pxelinux.0 pxeboot/menu.c32 \
    pxeboot/chain.c32 .../patchiso_build_XXXXXX/var/pxeboot/
  cp: target '.../patchiso_build_XXXXXX/var/pxeboot/' \
    is not a directory

This commit fixes the paths used by patch-iso so that such errors are
not encountered.

Verification:
- patch-iso is successfully executed against a patch file that includes
  a newer pxe-network-installer package.

Closes-Bug: 1968566

Change-Id: I4606d062b45ed03b64f514131d3d6e27434e7675
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>