debrepack.py: don't clean up the "srcdir" in setup
The debian folder is parsed by setup(), which is called by
download(), checksum(), package(), and only package() access the
"srcdir".

At beginning of package(), the "packdir" which is the upper directory of
"srcdir" is recreated, so don't need clean up "srcdir" in setup.

Test Plan:

Pass: build-pkgs -a

Story: 2008846
Task: 45052

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: I4a877c5325b35b85f77bbf9d764e26368ddd4ef2